Move Ergvein.Wallet.Input to Ergvein.Wallet.Elements.Input
